backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-generativeai>=0.8.0
pydantic==2.7.1
python-multipart==0.0.9
